requirements.txt
streamlit>=1.32.0
openai>=1.14.0
faiss-cpu>=1.8.0
numpy>=1.26.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain_community>=0.4.1
langsmith>=0.1.0
opensearch-py>=2.4.0
